generic update commit

+++ segmentation/views.py
@@ -2311,7 +2311,7 @@ def segmentation_handler(doc: bokeh.document.Document) -> None:
         hist, edges = np.histogram(n_osc_dk, bins=max_osc+2, range=(0, max_osc+2))
         source_nosc_dk.data={'x': edges[:-1], 'top': hist}
 
-        hist, edges = np.histogram(tod, bins=(max(tod, default=100.)-min(tod, default=0.))/30., range=(min(tod, default=0), max(tod, default=100.)))
+        hist, edges = np.histogram(tod, bins=int((max(tod, default=100.)-min(tod, default=0.))/30.), range=(min(tod, default=0), max(tod, default=100.)))
         source_tod.data={'x': edges[:-1], 'top': hist}
 
         hist, edges = np.histogram(start_osc, bins=nframes*10, range=(0, nframes*10))
